feat: 리뷰 기능 및 테스트 코드 추가

+++ service/review/src/main/java/innercircle/commerce/review/entity/Review.java
@@ -21,7 +21,7 @@ public class Review {
     private Long memberId;
     private Long productId;
 
-    private Long rating = 0L;
+    private int rating = 0;
     @Column(columnDefinition = "TEXT")
     private String content;
     private Long likeCount;
@@ -32,7 +32,7 @@ public class Review {
     private List<ReviewLike> likes = new ArrayList<>();
 
     @Builder
-    public Review(Long memberId, Long productId, long rating, String content) {
+    public Review(Long memberId, Long productId, int rating, String content) {
         this.memberId = memberId;
         this.productId = productId;
         this.rating = rating;
